[Skeleton] Move from lab to core
