+++ docs/src/pages.js
@@ -78,6 +78,7 @@ const pages = [
           { pathname: '/components/backdrop' },
           { pathname: '/components/dialogs' },
           { pathname: '/components/progress' },
+          { pathname: '/components/skeleton' },
           { pathname: '/components/snackbars' },
         ],
       },
@@ -121,6 +122,7 @@ const pages = [
           { pathname: '/components/autocomplete' },
           { pathname: '/components/pagination' },
           { pathname: '/components/skeleton' },
+          { pathname: '/components/rating' },
           { pathname: '/components/slider-styled' },
           { pathname: '/components/speed-dial' },
           { pathname: '/components/timeline' },
