+++ docs/src/pages/components/skeleton/SkeletonTypography.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import Typography from '@material-ui/core/Typography';
-import Skeleton from '@material-ui/lab/Skeleton';
+import Skeleton from '@material-ui/core/Skeleton';
 import Grid from '@material-ui/core/Grid';
 
 const variants = ['h1', 'h3', 'body1', 'caption'];
