+++ docs/src/pages/components/skeleton/SkeletonTypography.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Typography, { TypographyProps } from '@material-ui/core/Typography';
-import Skeleton from '@material-ui/lab/Skeleton';
+import Skeleton from '@material-ui/core/Skeleton';
 import Grid from '@material-ui/core/Grid';
 
 const variants = [
