+++ docs/src/pages/components/skeleton/Variants.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Skeleton from '@material-ui/lab/Skeleton';
+import Skeleton from '@material-ui/core/Skeleton';
 
 export default function Variants() {
   return (
