+++ packages/material-ui-lab/src/themeAugmentation/props.d.ts
@@ -2,7 +2,6 @@ import { AutocompleteProps } from '../Autocomplete';
 import { AvatarGroupProps } from '../AvatarGroup';
 import { PaginationProps } from '../Pagination';
 import { PaginationItemProps } from '../PaginationItem';
-import { SkeletonProps } from '../Skeleton';
 import { SpeedDialProps } from '../SpeedDial';
 import { SpeedDialActionProps } from '../SpeedDialAction';
 import { SpeedDialIconProps } from '../SpeedDialIcon';
@@ -25,7 +24,6 @@ export interface LabComponentsPropsList {
   MuiAvatarGroup: AvatarGroupProps;
   MuiPagination: PaginationProps;
   MuiPaginationItem: PaginationItemProps;
-  MuiSkeleton: SkeletonProps;
   MuiSpeedDial: SpeedDialProps;
   MuiSpeedDialAction: SpeedDialActionProps;
   MuiSpeedDialIcon: SpeedDialIconProps;
