+++ packages/material-ui/src/Skeleton/Skeleton.test.js
