+++ packages/material-ui/src/Skeleton/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './Skeleton';
+export * from './Skeleton';
