+++ packages/material-ui/src/Skeleton/index.js
@@ -1,0 +1,1 @@
+export { default } from './Skeleton';
