+++ packages/material-ui/src/styles/overrides.d.ts
@@ -1,4 +1,8 @@
 import { CSSProperties, StyleRules } from './withStyles';
+import { AccordionActionsClassKey } from '../AccordionActions';
+import { AccordionClassKey } from '../Accordion';
+import { AccordionDetailsClassKey } from '../AccordionDetails';
+import { AccordionSummaryClassKey } from '../AccordionSummary';
 import { AlertClassKey } from '../Alert';
 import { AlertTitleClassKey } from '../AlertTitle';
 import { AppBarClassKey } from '../AppBar';
@@ -29,10 +33,6 @@ import { DialogContentTextClassKey } from '../DialogContentText';
 import { DialogTitleClassKey } from '../DialogTitle';
 import { DividerClassKey } from '../Divider';
 import { DrawerClassKey } from '../Drawer';
-import { AccordionActionsClassKey } from '../AccordionActions';
-import { AccordionClassKey } from '../Accordion';
-import { AccordionDetailsClassKey } from '../AccordionDetails';
-import { AccordionSummaryClassKey } from '../AccordionSummary';
 import { FabClassKey } from '../Fab';
 import { FilledInputClassKey } from '../FilledInput';
 import { FormControlClassKey } from '../FormControl';
@@ -70,6 +70,7 @@ import { RadioClassKey } from '../Radio';
 import { RatingClassKey } from '../Rating';
 import { ScopedCssBaselineClassKey } from '../ScopedCssBaseline';
 import { SelectClassKey } from '../Select';
+import { SkeletonClassKey } from '../Skeleton';
 import { SliderClassKey } from '../Slider';
 import { SnackbarClassKey } from '../Snackbar';
 import { SnackbarContentClassKey } from '../SnackbarContent';
@@ -185,6 +186,7 @@ export interface ComponentNameToClassKey {
   MuiRating: RatingClassKey;
   MuiScopedCssBaseline: ScopedCssBaselineClassKey;
   MuiSelect: SelectClassKey;
+  MuiSkeleton: SkeletonClassKey;
   MuiSlider: SliderClassKey;
   MuiSnackbar: SnackbarClassKey;
   MuiSnackbarContent: SnackbarContentClassKey;
