+++ packages/material-ui/src/styles/props.d.ts
@@ -74,6 +74,7 @@ import { RadioProps } from '../Radio';
 import { RatingProps } from '../Rating';
 import { ScopedCssBaselineProps } from '../ScopedCssBaseline';
 import { SelectProps } from '../Select';
+import { SkeletonProps } from '../Skeleton';
 import { SliderProps } from '../Slider';
 import { SnackbarContentProps } from '../SnackbarContent';
 import { SnackbarProps } from '../Snackbar';
@@ -185,6 +186,7 @@ export interface ComponentsPropsList {
   MuiRating: RatingProps;
   MuiScopedCssBaseline: ScopedCssBaselineProps;
   MuiSelect: SelectProps;
+  MuiSkeleton: SkeletonProps;
   MuiSlider: SliderProps;
   MuiSnackbar: SnackbarProps;
   MuiSnackbarContent: SnackbarContentProps;
